# Conflicts:
#	src/main/java/com/github/unchama/seichiassist/effect/arrow/ArrowBlizzardTask.java
#	src/main/java/com/github/unchama/seichiassist/effect/arrow/ArrowMagicTask.java
#	src/main/java/com/github/unchama/seichiassist/effect/arrow/ArrowMeteoTask.java
#	src/main/java/com/github/unchama/seichiassist/listener/PlayerInventoryListener.kt

+++ src/main/java/com/github/unchama/seichiassist/task/GiganticBerserkTask.java
@@ -45,7 +45,7 @@ public class GiganticBerserkTask {
 
 			mana.increase(i,p, playerdata.getLevel());
 			player.sendMessage(ChatColor.YELLOW + "" + ChatColor.BOLD + "" + ChatColor.UNDERLINE + "Gigantic" + ChatColor.RED + "" + ChatColor.BOLD + "" + ChatColor.UNDERLINE + "Berserk" + ChatColor.WHITE + "の効果でマナが" + i +"回復しました");
-			player.playSound(player.getLocation(), Sound.ENTITY_WITHER_SHOOT, 1, (float)0.5) ;
+			player.playSound(player.getLocation(), Sound.ENTITY_WITHER_SHOOT, 1, 0.5f) ;
 		}
 
 		//最大レベルの場合終了
@@ -69,17 +69,17 @@ public class GiganticBerserkTask {
 				playerdata.setGBlevel(playerdata.getGBlevel() + 1);
 				//プレイヤーにメッセージ
 				player.sendMessage(ChatColor.YELLOW + "" + ChatColor.BOLD + "" + ChatColor.UNDERLINE + "Gigantic" + ChatColor.RED + "" + ChatColor.BOLD + "" + ChatColor.UNDERLINE + "Berserk" + ChatColor.WHITE + "のレベルがアップし、確率が上昇しました");
-				player.playSound(player.getLocation(), Sound.BLOCK_ENCHANTMENT_TABLE_USE, 1, (float)0.8) ;
+				player.playSound(player.getLocation(), Sound.BLOCK_ENCHANTMENT_TABLE_USE, 1, 0.8f) ;
 				//最大レベルになった時の処理
 				if(playerdata.getGBstage() == 4 && playerdata.getGBlevel() == 9){
-					Util.sendEverySound(Sound.ENTITY_ENDERDRAGON_DEATH, 1, (float)1.2);
+					Util.sendEverySound(Sound.ENTITY_ENDERDRAGON_DEATH, 1, 1.2f);
 					Util.sendEveryMessage(ChatColor.GOLD + "" + ChatColor.BOLD + playerdata.getName() + "がパッシブスキル:" + ChatColor.YELLOW + "" + ChatColor.BOLD + "" + ChatColor.UNDERLINE + "Gigantic" + ChatColor.RED + "" + ChatColor.BOLD + "" + ChatColor.UNDERLINE + "Berserk" + ChatColor.GOLD + "" + ChatColor.BOLD + "を完成させました！");
 				}
 			}
 			//レベルが10かつ段階がダイヤ未満の場合は進化待機状態へ
 			else if(playerdata.getGBstage() <= 3){
 				player.sendMessage(ChatColor.GREEN + "パッシブスキルメニューより" + ChatColor.YELLOW + "" + ChatColor.BOLD + "" + ChatColor.UNDERLINE + "Gigantic" + ChatColor.RED + "" + ChatColor.BOLD + "" + ChatColor.UNDERLINE + "Berserk" + ChatColor.GREEN + "スキルが進化可能です。");
-				player.playSound(player.getLocation(), Sound.BLOCK_ENCHANTMENT_TABLE_USE, 1, (float)0.8) ;
+				player.playSound(player.getLocation(), Sound.BLOCK_ENCHANTMENT_TABLE_USE, 1, 0.8f) ;
 				playerdata.setGBStageUp(true);
 			}
 		}
